+++ src/main/java/ru/karinkicks/dto/ProductDto.java
@@ -3,12 +3,14 @@ package ru.karinkicks.dto;
 import lombok.Data;
 import ru.karinkicks.entity.Product;
 
+import java.math.BigDecimal;
+
 
 @Data
 public class ProductDto {
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
 
     public ProductDto(Product product) {
         this.id = product.getId();
